src/requirements.txt
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
idna==3.10
numpy==2.2.4
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pycparser==2.22
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.1
reportlab==4.3.1
requests==2.32.3
six==1.17.0
tzdata==2025.1
urllib3==2.3.0
python-docx==1.1.0
